+++ entrypoints/match.content/sounds.ts
@@ -1,17 +1,16 @@
-import { AutodartsToolsConfig, AutodartsToolsCricketClosedPoints, AutodartsToolsMatchStatus } from "@/utils/storage";
+import { AutodartsToolsConfig, AutodartsToolsCricketClosedPoints, AutodartsToolsMatchStatus,AutodartsToolsGlobalStatus } from "@/utils/storage";
 import { AutodartsToolsCallerConfig } from "@/utils/callerStorage";
 import { AutodartsToolsSoundsConfig } from "@/utils/soundsStorage";
-
 import { playPointsSound, playSound } from "@/utils/playSound";
 import { isCricket, isValidGameMode } from "@/utils/helpers";
-
-let isFirstStartHandled = false; // Flag to track if the IsFirstStart event has been handled
+import { soundsPlayer } from "@/entrypoints/match.content/soundsPlayer";
 
 export async function sounds() {
   const config = await AutodartsToolsConfig.getValue();
   const isCallerEnabled = config.caller.enabled && isValidGameMode();
   const callerActive = (await AutodartsToolsCallerConfig.getValue()).caller.find(caller => caller.isActive);
 
+  const globalStatus = await AutodartsToolsGlobalStatus.getValue();
   const soundConfig = await AutodartsToolsSoundsConfig.getValue();
   const matchStatus = await AutodartsToolsMatchStatus.getValue();
   const cricketClosedPoints = await AutodartsToolsCricketClosedPoints.getValue();
@@ -30,19 +29,16 @@ export async function sounds() {
   const turnContainerEl = document.getElementById("ad-ext-turn");
   const letsGo = [...turnContainerEl?.querySelectorAll("div") as NodeListOf<HTMLElement>].filter(el => !el.classList.contains("ad-ext-turn-throw")).length === 4;
 
-  if (letsGo && isFirstStartHandled) {
-    if (!config.sounds.enabled) {
-      return;
-    } else if (soundConfig.playerStart?.data || !soundConfig.playerStart?.info) {  
-      return;
-    } else {
-      await playPlayerNameSound(callerServerUrl, callerFileExt, playerName, 1);
-   }
+  if (!globalStatus.isFirstStart && letsGo && config.sounds.enabled && (soundConfig.playerStart?.data || soundConfig.playerStart?.info)) {
+    const playerNameSoundPlayed = await soundsPlayer();
+    if (playerNameSoundPlayed) {
+      await new Promise(resolve => setTimeout(resolve, 1200));
+    }
+    await playSound("playerStart", 2);
   }
 
-  // Set isFirstStartHandled to true after the IsFirstStart event has occurred
-  isFirstStartHandled = true;
-
+  await AutodartsToolsGlobalStatus.setValue({ ...globalStatus, isFirstStart: false });
+  
   let curThrowPointsNumber: number = -1;
   let curThrowPointsBed: string = "";
   let curThrowPointsMultiplier: number = 1;
@@ -116,29 +112,4 @@ export async function sounds() {
       }
     }
   }, isBot ? 500 : 0);
-}
-
-async function playPlayerNameSound(callerServerUrl: string, callerFileExt: string, playerName: string | undefined, volume: number): Promise<void> {
-  if (!playerName) return; // No player name found
-
-  const soundUrl = `${callerServerUrl}${playerName.toLowerCase()}${callerFileExt}`;
-
-  try {
-    const response = await fetch(soundUrl, { method: 'HEAD' });
-    if (response.ok) {
-      const audio = new Audio(soundUrl);
-      setTimeout(() => {
-        playSound("playerStart");
-      }, 1200); // Delay the "playerStart" sound by 1.2 seconds
-
-      audio.volume = volume;
-      audio.play();
-    } else {
-      console.log(`The file ${soundUrl} does not exist.`);
-      playSound("playerStart"); // Play "playerStart" sound if player's name sound does not exist
-    }
-  } catch (error) {
-    console.error(`Error loading player name sound: ${error}`);
-    playSound("playerStart"); // Play "playerStart" sound if there is an error loading player's name sound
-  }
 }
